Move MQTT pub/sub class and transaction class to network package
