+++ src/main/java/jp/ex_t/kazuaki/change_vision/apply_transaction/ClassDiagramApplyTransaction.kt
@@ -14,6 +14,7 @@ import com.change_vision.jude.api.inf.model.*
 import com.change_vision.jude.api.inf.presentation.ILinkPresentation
 import com.change_vision.jude.api.inf.presentation.INodePresentation
 import jp.ex_t.kazuaki.change_vision.*
+import jp.ex_t.kazuaki.change_vision.network.*
 import java.awt.geom.Point2D
 
 class ClassDiagramApplyTransaction: IApplyTransaction {
