+++ src/main/java/jp/ex_t/kazuaki/change_vision/apply_transaction/ReflectTransaction.kt
@@ -13,7 +13,7 @@ import com.change_vision.jude.api.inf.exception.BadTransactionException
 import com.change_vision.jude.api.inf.project.ProjectAccessor
 import com.change_vision.jude.api.inf.ui.IPluginActionDelegate.UnExpectedException
 import jp.ex_t.kazuaki.change_vision.Logging
-import jp.ex_t.kazuaki.change_vision.Transaction
+import jp.ex_t.kazuaki.change_vision.network.Transaction
 import jp.ex_t.kazuaki.change_vision.event_listener.ProjectChangedListener
 import jp.ex_t.kazuaki.change_vision.logger
 import javax.swing.SwingUtilities
